chore: move RetryStage into its own file
